+++ NEWS.md
@@ -1,3 +1,12 @@
+# darksky 1.3.0
+
+* Remove dplyr dependency (replacing it with a plyr dep but that was implicitly there
+  by the ggplot2 dependency)
+
+# darkksy 1.2.0
+
+* `get_forcast_for()`'s `timestamp` parameter can now be a `POSIXct` or `Date` object and it will auto-convert to the proper API format
+
 # darksky 1.0.0
 
 * CRAN release